- 调整测试用例结构
- oo 模块名称变更为 class

- 升级 jasmine 版本
