+++ package.json
@@ -20,8 +20,8 @@
             "url": "https://github.com/xiaoxin0818"
         },
         {
-          "name": "yanghuabei",
-          "url": "https://github.com/yanghuabei"
+            "name": "yanghuabei",
+            "url": "https://github.com/yanghuabei"
         }
     ],
     "keywords": [
@@ -43,11 +43,12 @@
     },
     "homepage": "https://github.com/ecomfe/oo",
     "devDependencies": {
+        "jasmine-core": "^2.3.4",
         "karma": "^0.12.23",
         "karma-chrome-launcher": "^0.1.4",
         "karma-coverage": "^0.2.4",
         "karma-ie-launcher": "^0.1.5",
-        "karma-jasmine": "^0.2.2",
+        "karma-jasmine": "^0.3.6",
         "karma-phantomjs-launcher": "^0.1.4"
     },
     "edp": {
